+++ apps/igloo-kit-cli/src/infrastructure/setup/validate.rs
@@ -1,11 +1,11 @@
-use std::{fs, env, io::{self, Write}, error::Error};
+use std::io::{self, Write, Error};
 
 use crate::infrastructure::docker;
 
 
 // TODO: Add clickhouse validation
 
-pub fn validate_docker_run(debug: bool) -> Result<(), Box<dyn Error>> {
+pub fn validate_docker_run(debug: bool) -> Result<(), Error> {
     let output = docker::filter_list_containers("redpanda-1");
 
     match output {
@@ -19,17 +19,17 @@ pub fn validate_docker_run(debug: bool) -> Result<(), Box<dyn Error>> {
                 Ok(())
             } else {
                 println!("Failed to validate docker container");
-                Err(Box::new(io::Error::new(io::ErrorKind::Other, "Failed to validate red panda container exists")))
+                return Err(io::Error::new(io::ErrorKind::Other, "Failed to validate red panda container exists"))
             }
         },
         Err(err) => {
             println!("Failed to validate docker container");
-            Err(Box::new(err))
+            Err(err)
         },
     }
 }
 
-pub fn validate_panda_house_network(debug: bool) -> Result<(), Box<dyn Error>> {
+pub fn validate_panda_house_network(debug: bool) -> Result<(), Error> {
     let output = docker::network_list();
 
     match output {
@@ -43,18 +43,18 @@ pub fn validate_panda_house_network(debug: bool) -> Result<(), Box<dyn Error>> {
                 Ok(())
             } else {
                 println!("Failed to validate panda_house_network");
-                Err(Box::new(io::Error::new(io::ErrorKind::Other, "Failed to validate panda_house network")))
+                Err(io::Error::new(io::ErrorKind::Other, "Failed to validate panda_house network"))
             }
         },
         Err(err) => {
             println!("Failed to validate panda_house_network");
-            Err(Box::new(err))
+            Err(err)
         },
         
     }
 }
 
-pub fn validate_red_panda_cluster(debug: bool) -> Result<(), Box<dyn Error>> {
+pub fn validate_red_panda_cluster(debug: bool) -> Result<(),  Error> {
     let output = docker::run_rpk_list();
 
     match output {
@@ -68,13 +68,13 @@ pub fn validate_red_panda_cluster(debug: bool) -> Result<(), Box<dyn Error>> {
                 Ok(())
             } else {
                 println!("Failed to validate docker container");
-                Err(Box::new(io::Error::new(io::ErrorKind::Other, "Failed to validate red panda cluster")))
+                Err(io::Error::new(io::ErrorKind::Other, "Failed to validate red panda cluster"))
 
             }
         },
         Err(err) => {
             println!("Failed to validate redpanda cluster");
-            Err(Box::new(err))
+            Err(err)
         },
     }
 }
